feat: moving util files into /lib/ and fixing imports
