+++ exporter/SynthesisFusionAddin/src/UI/ConfigCommand.py
@@ -26,8 +26,8 @@ from src.Parser.ExporterOptions import ExporterOptions
 from src.Parser.SynthesisParser.Utilities import guid_occurrence
 from src.Types import SELECTABLE_JOINT_TYPES, ExportLocation, ExportMode
 from src.UI import FileDialogConfig
-from src.UI.Handlers import PersistentEventHandler
-from src.Util import convertMassUnitsTo, designMassCalculation
+from src.lib.Handlers import PersistentEventHandler
+from src.lib.Util import convertMassUnitsTo, designMassCalculation
 
 generalConfigTab: GeneralConfigTab.GeneralConfigTab
 jointConfigTab: JointConfigTab.JointConfigTab
@@ -292,6 +292,8 @@ class IncomingHTMLMessageHandler(PersistentEventHandler, adsk.core.HTMLEventHand
         elif html_args.action == "cancelSelection":
             gm.ui.terminateActiveCommand()
             html_args.returnData = "{}"
+        elif html_args.action == "getDesignChecks":
+            
         else:
             gm.ui.messageBox(f"Event {html_args.action} arrived<span>{json.dumps(data, indent=2)}</span>")
