+++ exporter/SynthesisFusionAddin/src/UI/TaggingConfigTab.py
@@ -2,7 +2,7 @@ import adsk.core
 import adsk.fusion
 
 from src.Logging import getLogger, logFailure
-from src.UI import IconPaths
+from src.lib import IconPaths
 from src.UI.CreateCommandInputsHelper import createTableInput, createTextBoxInput
 
 logger = getLogger()
