+++ exporter/SynthesisFusionAddin/src/lib/GlobalManager.py
