+++ exporter/SynthesisFusionAddin/src/lib/Helper.py
@@ -4,7 +4,8 @@ import adsk.core
 
 from src import APP_NAME, APP_TITLE, INTERNAL_ID, gm
 from src.Logging import logFailure
-from src.UI import HUI, Events
+from src.UI import HUI
+from src.lib import Events
 
 
 def check_solid_open() -> bool:
